chore: move folders and use ts paths
