+++ frontend/package.json
@@ -16,6 +16,7 @@
     "@nuxt/ui": "3.0.0-alpha.11",
     "@tailwindcss/postcss": "^4.0.6",
     "@tailwindcss/vite": "^4.0.6",
+    "@types/node": "^22.13.4",
     "@unhead/vue": "^1.11.19",
     "@vitejs/plugin-vue": "^5.2.1",
     "@vueuse/core": "^12.7.0",
