+++ frontend/src/components/AppBreadcrumb.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref, watchEffect } from "vue";
-import { useUrlParams } from "../composables/useUrlParams";
-import { useConnections } from "../composables/useConnections";
+import { useUrlParams } from "@/composables/useUrlParams";
+import { useConnections } from "@/composables/useConnections";
 
 const { databaseId, schemaId, tableId } = useUrlParams();
 const { connections } = useConnections();
