+++ frontend/src/components/database/AppSchema.vue
@@ -1,13 +1,16 @@
 <script setup lang="ts">
 import { useRouter } from "vue-router";
-import { useUrlParams } from "../../composables/useUrlParams";
-import { FormattedQueryResult, RowAction } from "./table/table";
-import { useWails } from "../../wails";
-import { GetSchemaTables } from "../../../wailsjs/go/app/App";
+import { useUrlParams } from "@/composables/useUrlParams";
+import {
+  FormattedQueryResult,
+  RowAction,
+} from "@/components/database/table/table";
+import { useWails } from "@/composables/useWails";
+import { GetSchemaTables } from "_/go/app/App";
 import { Effect } from "effect";
 import { ref } from "vue";
-import { client } from "../../../wailsjs/go/models";
-import { formatQueryResult } from "../../effects/columns";
+import { client } from "_/go/models";
+import { formatQueryResult } from "@/effects/columns";
 
 const router = useRouter();
 const { databaseId, schemaId } = useUrlParams();
