+++ frontend/src/components/database/table/cell/AppTypeSelect.vue
