+++ frontend/vite.config.ts
@@ -3,8 +3,15 @@ import vue from "@vitejs/plugin-vue";
 import tailwindcss from "@tailwindcss/vite";
 import ui from "@nuxt/ui/vite";
 import packageJson from "./package.json";
+import { resolve } from "path";
 
 export default defineConfig({
+  resolve: {
+    alias: [
+      { find: "@", replacement: resolve(__dirname, "src") },
+      { find: "_", replacement: resolve(__dirname, "wailsjs") },
+    ],
+  },
   define: {
     "import.meta.env.PACKAGE_VERSION": JSON.stringify(packageJson.version),
   },
